+++ 07.bowling_object/shot.rb
@@ -1,1 +1,13 @@
 # frozen_string_literal: true
+
+class Shot
+  attr_reader :mark
+
+  def initialize(mark)
+    @mark = mark
+  end
+
+  def score
+    @mark == 'X' ? 10 : @mark.to_i
+  end
+end
